GroupMove/Entity Loads
Groupunit movement has been updated to move with less errors, but a
little bit less organic behavior. It is still unfinished, but it is good
enough to move on with the other critical features for the Nov 7th demo.
SetFacingDirection() now returns a float, the difference between the
current facing direction and where we want to face (Y-Axis). It is used
to slow the character down on sharp turns.

As part of the error fixing on the GroupManager Lost Unit
functionalities, which is no longer currently in use, the Entity
components of characters are now correctly added in the LoadStats method
on controllers. The Method is currently public as part of the interface,
but it is unlikely that this is necessary or wise. Will likely remove it
from the interface.

Minor tweaks all around.

Player DirectionalMovement can still sometimes result in the player grid
unit being stuck in obstalces.

The Entity system was overhauled, No longer is a Monobehavior, so as to
allow the use of constructors. At the base level Entity Class, an
Entity_MonoBehaviour object is added to the entity's gameobject, with
said Entity as the "Owner" of that Unity-Component. This preserves the
ability to find and damage Entity using unity physics functions.

PlayerEntity and OrkEntity (isa TaggableEntity) created. Currently only
printing out the type of death, will add appropriately unique functions
later.

OrkMelee Prefab added
OrkGroup Prefab added

various project script folders restructed